+++ waldiez/__main__.py
@@ -1,6 +1,6 @@
 """Waldiez entrypoint when called as a module."""
 
-from .cli import main
+from .cli import app
 
 if __name__ == "__main__":
-    main()
+    app()